# frontend/src/main/resources/openapi.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.0.0",
    "title" : "JurassicCRM",
    "termsOfService" : "Terms",
    "contact" : {
      "name" : "Dmitrii Fitisov",
      "url" : "https://google.com",
      "email" : "fitisovdmtr@gmail.com"
    }
  },
  "host" : "http://localhost",
...
  "schemes" : [ "http" ],
  "paths" : {
...
    "/api/aviary/type/{id}" : {
      "put" : {
        "tags" : [ "aviaryType" ],
        "summary" : "updateAviary",
        "description" : "",
        "operationId" : "updateAviary",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/SimpleEntityInputTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SimpleEntityOutputTO"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "aviaryType" ],
        "summary" : "deleteAviary",
        "description" : "",
        "operationId" : "deleteAviary",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SimpleEntityOutputTO"
            }
          }
        }
      }
    },
    "/api/decoration/type" : {
      "get" : {
        "tags" : [ "decorationType" ],
        "summary" : "getAllDecorations",
        "description" : "",
        "operationId" : "getAllDecorations",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/SimpleEntityOutputTO"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "decorationType" ],
        "summary" : "createDecoration",
        "description" : "",
        "operationId" : "createDecoration",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/SimpleEntityInputTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SimpleEntityOutputTO"
            }
          }
        }
      }
    },
    "/api/decoration/type/{id}" : {
      "put" : {
        "tags" : [ "decorationType" ],
        "summary" : "updateDecoration",
        "description" : "",
        "operationId" : "updateDecoration",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/SimpleEntityInputTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SimpleEntityOutputTO"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "decorationType" ],
        "summary" : "deleteDecoration",
        "description" : "",
        "operationId" : "deleteDecoration",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SimpleEntityOutputTO"
            }
          }
        }
      }
    },
...
    "/api/dinosaur/type/{id}" : {
      "put" : {
        "tags" : [ "dinosaur" ],
        "summary" : "updateDinosaurType",
        "description" : "",
        "operationId" : "updateDinosaurType",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/SimpleEntityInputTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SimpleEntityOutputTO"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "dinosaur" ],
        "summary" : "deleteDinosaurType",
        "description" : "",
        "operationId" : "deleteDinosaurType",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SimpleEntityOutputTO"
            }
          }
        }
      }
    },
    "/api/document/RESEARCH_DATA" : {
      "post" : {
        "tags" : [ "document" ],
        "summary" : "createResearchData",
        "description" : "",
        "operationId" : "createResearchData",
        "consumes" : [ "multipart/form-data" ],
        "parameters" : [ {
          "name" : "researchData",
          "in" : "formData",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "attachment",
          "in" : "formData",
          "required" : true,
          "type" : "file"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/DocumentOutputTO"
            }
          }
        }
      }
    },
...
    "/api/document/{documentType}" : {
      "get" : {
        "tags" : [ "document" ],
        "summary" : "getDocumentsByType",
        "description" : "",
        "operationId" : "getDocumentsByType",
        "parameters" : [ {
          "name" : "documentType",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "enum" : [ "THEME_ZONE_PROJECT", "DINOSAUR_PASSPORT", "TECHNOLOGICAL_MAP", "AVIARY_PASSPORT", "RESEARCH_DATA" ]
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/DocumentOutputTO"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "document" ],
        "summary" : "createDocument",
        "description" : "",
        "operationId" : "createDocument",
        "consumes" : [ "multipart/form-data" ],
        "parameters" : [ {
          "name" : "documentType",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "enum" : [ "THEME_ZONE_PROJECT", "DINOSAUR_PASSPORT", "TECHNOLOGICAL_MAP", "AVIARY_PASSPORT", "RESEARCH_DATA" ]
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/DocumentOutputTO"
            }
          }
        }
      }
    },
    "/api/document/{documentType}/{documentId}" : {
      "put" : {
        "tags" : [ "document" ],
        "summary" : "updateDocument",
        "description" : "",
        "operationId" : "updateDocument",
        "parameters" : [ {
          "name" : "documentType",
          "in" : "path",
          "required" : true,
          "type" : "string",
          "enum" : [ "THEME_ZONE_PROJECT", "DINOSAUR_PASSPORT", "TECHNOLOGICAL_MAP", "AVIARY_PASSPORT", "RESEARCH_DATA" ]
        }, {
          "name" : "documentId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/DocumentOutputTO"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "DocumentOutputTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string",
          "enum" : [ "THEME_ZONE_PROJECT", "DINOSAUR_PASSPORT", "TECHNOLOGICAL_MAP", "AVIARY_PASSPORT", "RESEARCH_DATA" ]
        },
        "author" : {
          "$ref" : "#/definitions/UserOutputTO"
        },
        "lastUpdater" : {
          "$ref" : "#/definitions/UserOutputTO"
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        },
        "lastUpdate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "description" : {
          "type" : "string"
        }
      }
    },
...
    "SimpleEntityInputTO" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        }
      }
    },
    "SimpleEntityOutputTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        }
      }
    },
...
    "UserOutputTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "username" : {
          "type" : "string"
        },
        "firstName" : {
          "type" : "string"
        },
        "lastName" : {
          "type" : "string"
        },
        "department" : {
          "type" : "string",
          "enum" : [ "RESEARCH", "INCUBATION", "SECURITY", "ADMINISTRATION", "MAINTENANCE", "ACCOMMODATION" ]
        }
      }
    }
  }
}